requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.2.0
drf-excel==2.1.0
drf-renderer-xlsx==1.0.0
et-xmlfile==1.1.0
openpyxl==3.0.10
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
unicodecsv==0.14.1
psycopg2-binary==2.8.6

